feat: 경매 입찰을 Auctioneer 클래스를 통해 진행한다.
