+++ src/main/java/com/wootecam/luckyvickyauction/core/auction/controller/BuyerAuctionController.java
@@ -8,7 +8,6 @@ import com.wootecam.luckyvickyauction.core.auction.service.AuctionService;
 import com.wootecam.luckyvickyauction.core.member.controller.BuyerOnly;
 import com.wootecam.luckyvickyauction.core.member.controller.Login;
 import com.wootecam.luckyvickyauction.core.member.dto.SignInInfo;
-import com.wootecam.luckyvickyauction.core.payment.service.PaymentProxyService;
 import com.wootecam.luckyvickyauction.core.payment.service.PaymentService;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -27,7 +26,6 @@ public class BuyerAuctionController {
 
     private final AuctionService auctionService;
     private final PaymentService paymentService;
-    private final PaymentProxyService paymentProxy;
 
     // 사용자는 경매 목록을 조회한다.
     @GetMapping("/auctions")
@@ -50,7 +48,7 @@ public class BuyerAuctionController {
                                               @CurrentTime LocalDateTime now,
                                               @PathVariable(name = "auctionId") Long auctionId,
                                               @RequestBody PurchaseRequest purchaseRequest) {
-        paymentProxy.process(signInInfo, purchaseRequest.price(), auctionId, purchaseRequest.quantity(), now);
+        paymentService.process(signInInfo, purchaseRequest.price(), auctionId, purchaseRequest.quantity(), now);
         return ResponseEntity.ok().build();
     }
